Move invitations to background task
